Some UI refactorings
